move all htpasswd stuff into basic/__init__.py, delete unneeded class -> subclass
